Create output directory to avoid it being owned by root
If the output directory doesn't exist, docker will create it for us. Unfortunately,
this happens before docker switches to the --user specified. So we end up with a
directory owned by root that we can't write into.
Sidestep this issue by making sure the directory exists in the wrapper script.

Signed-off-by: Kai Blin <kblin@biosustain.dtu.dk>